Move `ActionsEnvVars` to `environment.ts`
